Add audio monitor and LEDFX autopause
Basically, I wanted to monitor my default outpt device and make
sure that LEDFX wasn't blanking my pretty WLED displays when
my audio player was off. I was originally going to do an
integration with spotify (and I still might) that watched if there
was a song on deck, but I decided to monitor the audio device
instead in order to avoid a separate OAuth2 authorization app/step,
and also so that the system would still work with movies or other
player apps.